Refrain from calling delete on session objects without ID.
git-svn-id: svn+ssh://svn.bmc.uu.se/cbe/openexam-php/trunk@2262 58b8aac4-d633-40de-91bf-eab09b0297ab

+++ version2/phalcon-mvc/app/library/Database/SessionAdapter.php
@@ -97,13 +97,15 @@ class SessionAdapter extends AdapterBase implements AdapterInterface
         public function write($sessionId, $data)
         {
                 if (!$this->session) {
-                        if (($this->session = SessionModel::findFirstBySessionId($sessionId)) == false) {
-                                $this->session = new SessionModel();
-                        }
+                        $this->session = SessionModel::findFirstBySessionId($sessionId);
                 }
-                if (empty($data)) {
-                        $this->session->delete();
+
+                if (empty($data) && $this->session) {
+                        return $this->session->delete();
+                } elseif (empty($data)) {
                         return true;
+                } elseif (!$this->session) {
+                        $this->session = new SessionModel();
                 }
 
                 if ($this->session->data == $data) {
